Refactor InterviewIDE with modular hooks and services
